+++ src/adchelper.cpp
@@ -20,7 +20,7 @@ ADCHelper::ADCHelper(uint8_t channels, float vref, uint8_t resolution, uint16_t 
   }
 }
 
-unsigned int ADCHelper::getNewReading(uint8_t channel)
+uint32_t ADCHelper::getNewReading(uint8_t channel)
 {
   // save the old one.
   uint8_t oldChannel = _currentChannel;
@@ -40,7 +40,7 @@ unsigned int ADCHelper::getNewReading(uint8_t channel)
     vTaskDelay(1);
 
   // okay load it in.
-  unsigned int reading = loadReading(channel);
+  uint32_t reading = loadReading(channel);
 
   // resume our regular scrolling
   requestReading(oldChannel);
@@ -55,9 +55,9 @@ float ADCHelper::getNewVoltage(uint8_t channel)
 // dummy functions - should be implemented by the real clases.
 void ADCHelper::requestADCReading(uint8_t channel) {}
 bool ADCHelper::isADCReady() { return false; }
-unsigned int ADCHelper::loadReadingFromADC(uint8_t channel) { return 0; }
+uint32_t ADCHelper::loadReadingFromADC(uint8_t channel) { return 0; }
 
-unsigned int ADCHelper::getReadingCount(uint8_t channel)
+uint16_t ADCHelper::getReadingCount(uint8_t channel)
 {
   return _averages[channel].count();
 }
@@ -74,10 +74,10 @@ void ADCHelper::printDebug()
     Serial.printf("%d Channels | Vref: %.3f | Resolution: %d\n", _totalChannels, _vref, _resolution);
 
     for (byte i = 0; i < _totalChannels; i++) {
-      unsigned int cnt = getReadingCount(i);
+      uint16_t cnt = getReadingCount(i);
       size_t cap = _averages[i].cap();
       uint32_t window = _averages[i].window();
-      unsigned int avgr = getAverageReading(i);
+      uint32_t avgr = getAverageReading(i);
       float avgv = getAverageVoltage(i);
       Serial.printf("CH%d: Window: %dms | Readings: %d/%d | Average: %d | Voltage: %.3f\n", i, window, cnt, cap, avgr, avgv);
     }
@@ -86,21 +86,21 @@ void ADCHelper::printDebug()
   }
 }
 
-unsigned int ADCHelper::getLatestReading(uint8_t channel)
+uint32_t ADCHelper::getLatestReading(uint8_t channel)
 {
   if (channel >= _totalChannels)
     return 0;
   return _averages[channel].latest();
 }
 
-unsigned int ADCHelper::getAverageReading(uint8_t channel)
+uint32_t ADCHelper::getAverageReading(uint8_t channel)
 {
   if (channel >= _totalChannels)
     return 0;
   return _averages[channel].average();
 }
 
-float ADCHelper::toVoltage(unsigned int reading)
+float ADCHelper::toVoltage(uint32_t reading)
 {
   // DUMP(reading);
   // DUMP(this->vref);
@@ -131,9 +131,9 @@ bool ADCHelper::isReady()
 }
 bool ADCHelper::isBusy() { return _isBusy; }
 
-unsigned int ADCHelper::loadReading(uint8_t channel)
+uint32_t ADCHelper::loadReading(uint8_t channel)
 {
-  unsigned int reading = loadReadingFromADC(channel);
+  uint32_t reading = loadReadingFromADC(channel);
 
   if (channel < _totalChannels)
     _averages[channel].add(reading);
@@ -181,7 +181,7 @@ ADS1115Helper::ADS1115Helper(float vref, ADS1115* adc, uint16_t samples, uint32_
 void ADS1115Helper::requestADCReading(uint8_t channel) { _adc->requestADC(channel); }
 bool ADS1115Helper::isADCReady() { return _adc->isReady(); }
 
-unsigned int ADS1115Helper::loadReadingFromADC(uint8_t channel)
+uint32_t ADS1115Helper::loadReadingFromADC(uint8_t channel)
 {
   int16_t reading;
   reading = _adc->getValue();
@@ -217,7 +217,7 @@ bool MCP3425Helper::isADCReady()
   return status.isReady();
 }
 
-unsigned int MCP3425Helper::loadReadingFromADC(uint8_t channel)
+uint32_t MCP3425Helper::loadReadingFromADC(uint8_t channel)
 {
   long value = 0;
   MCP342x::Config status;
@@ -251,12 +251,12 @@ bool MCP3564Helper::isADCReady()
   // return this->_adc->isComplete();
 }
 
-unsigned int MCP3564Helper::loadReadingFromADC(uint8_t channel)
+uint32_t MCP3564Helper::loadReadingFromADC(uint8_t channel)
 {
   return this->_adc->analogRead(_channelAddresses[channel]);
 }
 
-// float MCP3564Helper::toVoltage(unsigned int reading)
+// float MCP3564Helper::toVoltage(uint32_t reading)
 // {
 //   return reading * this->adc->getReference() / (this->adc->getMaxValue() / 2);
 // }
